// client/models/ReservationResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ReservationResponseDto = {
    id: string;
    customer: Record<string, any>;
    datetime: string;
    notes?: string;
    markAsDone: boolean;
    isActive: boolean;
    createdAt: string;
    updatedAt: string;
};